agent/requirements.txt
langchain==1.2.13
langchain-google-genai==1.0.0
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langsmith==0.7.22
web3==7.14.1
eth-account==0.13.7
python-telegram-bot==22.7
python-dotenv==1.2.2
schedule==1.2.2
requests==2.32.5
google-generativeai==0.6.0
pydantic==2.12.5
